fix: address PRRT_kwDOSJAM6s6bXiqn — break verdict reference import cycle
Move shared CommonMark label primitives into helpers_verdict_markdown and
defer emphasis's reference imports to runtime so helpers_verdict_reference
can load without a circular dependency on helpers_verdict_emphasis.
